requirements.txt
# NationBot Dependencies
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.6

# AI/ML
google-generativeai==0.3.2
tiktoken==0.5.2

# Vector DB (pgvector)
pgvector==0.2.4

# Graph Analysis
networkx==3.2.1

# HTTP & Parsing
httpx==0.26.0
feedparser==6.0.10
beautifulsoup4==4.12.3

# SSE
sse-starlette==1.8.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Testing
pytest-asyncio==0.23.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
